+++ application/GlkTextBufferWindow.h
@@ -19,8 +19,8 @@
 
 - (id) initWithContainerSize: (NSSize)size;
 - (void) clearImages;
-- (void) addImage: (NSImage*)image align: (int)align at: (int)top size: (NSSize)size;
-- (void) flowBreakAt: (int)pos;
+- (void) addImage: (NSImage*)image align: (NSInteger)align at: (NSInteger)top size: (NSSize)size;
+- (void) flowBreakAt: (NSInteger)pos;
 - (void) drawRect: (NSRect)rect;
 - (void) invalidateLayout;
 
@@ -40,20 +40,20 @@
     MarginContainer *container;
     NSTextView *textview;
     
-    int char_request;
-    int line_request;
-    int fence;		/* for input line editing */
-    int lastseen;	/* for more paging */
-    int lastchar;	/* for smart formatting */
+    NSInteger char_request;
+    NSInteger line_request;
+    NSInteger fence;		/* for input line editing */
+    NSInteger lastseen;	/* for more paging */
+    NSInteger lastchar;	/* for smart formatting */
     
     NSString *history[HISTORYLEN];
-    int historypos;
-    int historyfirst, historypresent;    
+    NSInteger historypos;
+    NSInteger historyfirst, historypresent;
 }
 
 - (void) recalcBackground;
 - (void) onKeyDown: (NSEvent*)evt;
-- (int) lastchar;
+- (NSInteger) lastchar;
 
 @end
